Refactor: theme-settings unit tests into separate suites
